+++ src/fidesops/ops/schemas/saas/saas_config.py
@@ -263,6 +263,7 @@ class SaaSType(Enum):
     sendgrid = "sendgrid"
     datadog = "datadog"
     rollbar = "rollbar"
+    braze = "braze"
 
 
 class SaaSConfigBase(BaseModel):